requirements.txt
airtest==1.2.7
pocoui==1.0.88
pocounit==1.0.40
loguru==0.6.0
pytest==7.2.0
Faker==15.1.3